Add files via upload
The tables presents the scores assigned by participants P1-P3 expert (XR), P4-P6 expert (HCI) and P7-P9 stakeholder. The overall SUS score values are M=68.05 with  SD=16.23. Considering the types of participants are expert (XR) 𝑀 = 56.66 with 𝑆𝐷 = 15.27, expert (HCI) 𝑀 = 68.33 with 𝑆𝐷 = 20.05 and stakeholder 𝑀 = 79.16 with 𝑆𝐷 = 6.29. The devices are HMD 𝑀 = 70.83 with 𝑆𝐷 = 26.73, notebook 𝑀 = 73.33 with 𝑆𝐷 = 5.77, and smartphone 𝑀 = 60 with 𝑆𝐷 = 6.29.